requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents[otel]
strands-agents-tools
aws-opentelemetry-distro>=0.10.0
boto3
streamlit
python-dotenv